+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.1] - 2020-04-12
+
+### Added
+
+- `.clang-tidy` at the top level, instead of calling arguments with CMake.
+- `VSCGEStaticAnalyzers.cmake` to facilitate module-level static checking.
+
+### Changed
+
+- `misc` module is now a global include directory.
+- tests now are done at the module level.
+
 ## [0.14.0] - 2020-04-05
 
 ### Added